# PERF: don't sort data twice in groupby apply when not using libreduction fast_apply