@jreback 2021-03-02
cool. can you add a whatsnew note about perf :->