@jreback 2021-03-02
thanks @jorisvandenbossche (prob with future PRs can just add the issue numbers onto this perf for group note, but up 2 you)